profcheck: refactor code
Refactor code so it can be used as standalone executable but also imported as dependency to run the validation.
